+++ crates/merman-render/src/svg/parity/class/mod.rs
@@ -31,11 +31,12 @@ mod groups;
 mod interface;
 
 mod label;
-use label::class_apply_inline_styles;
 
 mod namespace;
 
 mod node;
+
+mod nodes;
 
 mod note;